Review fixes

+++ irohad/model/converters/impl/pb_common.cpp
@@ -21,8 +21,7 @@ namespace iroha {
   namespace model {
     namespace converters {
 
-      protocol::Amount serializeAmount(
-          iroha::Amount iroha_amount) {
+      protocol::Amount serializeAmount(iroha::Amount iroha_amount) {
         protocol::Amount res;
         res.set_precision(iroha_amount.getPrecision());
         auto value = res.mutable_value();
@@ -34,12 +33,11 @@ namespace iroha {
         return res;
       }
 
-      iroha::Amount deserializeAmount(
-          protocol::Amount pb_amount) {
+      iroha::Amount deserializeAmount(protocol::Amount pb_amount) {
         auto value = pb_amount.value();
         return {value.first(), value.second(), value.third(), value.fourth(),
-                pb_amount.precision()};
+                static_cast<uint8_t>(pb_amount.precision())};
       }
-    }
-  }
-}
+    }  // namespace converters
+  }    // namespace model
+}  // namespace iroha
